move renderToast to utility file
